improved upload handling, next thing is fixing the fact that alignments are not shifted properly, it seems
